requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorating==0.6.1
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mal-api==0.4.0
MarkupSafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
Werkzeug==2.3.7
